# [SPARK-37099][SQL] Introduce the group limit of Window for rank-based filter to optimize top-k computation