--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/dsl/package.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/dsl/package.scala
@@ -467,13 +467,6 @@ package object dsl {
           orderSpec: Seq[SortOrder]): LogicalPlan =
         Window(windowExpressions, partitionSpec, orderSpec, logicalPlan)
 
-      def windowGroupLimit(
-          partitionSpec: Seq[Expression],
-          orderSpec: Seq[SortOrder],
-          rankLikeFunction: Expression,
-          limit: Int): LogicalPlan =
-        WindowGroupLimit(partitionSpec, orderSpec, rankLikeFunction, limit, logicalPlan)
-
       // TODO: Remove at Spark 4.0.0
       @deprecated("Use subquery(alias: String)", "3.4.0")
       def subquery(alias: Symbol): LogicalPlan = SubqueryAlias(alias.name, logicalPlan)