--- a/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicLogicalOperators.scala
+++ b/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/basicLogicalOperators.scala
@@ -1235,7 +1235,7 @@ case class WindowGroupLimit(
     orderSpec: Seq[SortOrder],
     rankLikeFunction: Expression,
     limit: Int,
-    child: LogicalPlan) extends OrderPreservingUnaryNode {
+    child: LogicalPlan) extends UnaryNode {
   assert(orderSpec.nonEmpty && limit > 0)
 
   override def output: Seq[Attribute] = child.output