… sql/catalyst/src/main/scala/org/apache/spark/sql/internal/SQLConf.scala
… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/InsertWindowGroupLimitSuite.scala
… sql/core/benchmarks/TopKBenchmark-results.txt
… sql/core/src/main/scala/org/apache/spark/sql/execution/QueryExecution.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/RemoveRedundantWindowGroupLimits.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/SparkOptimizer.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/SparkStrategies.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/adaptive/AdaptiveSparkPlanExec.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/window/WindowGroupLimitExec.scala
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q44/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q67.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q67.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q67/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q67/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q70.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q70.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q70/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q70/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q67a.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q67a.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q67a/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q67a/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q70a.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q70a.sf100/simplified.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q70a/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v2_7/q70a/simplified.txt
… sql/core/src/test/scala/org/apache/spark/sql/DataFrameWindowFunctionsSuite.scala