### What changes were proposed in this pull request?
introduce a new node `WindowGroupLimit` to filter out unnecessary rows based on rank computed on partial dataset.

it supports following pattern:
```
SELECT (... (row_number|rank|dense_rank)()
    OVER (
PARTITION BY ...
ORDER BY  ... ) AS rn)
WHERE rn (==|<|<=) k
        AND other conditions
```

For these three rank-like functions (row_number|rank|dense_rank), the rank of a key computed on partial dataset always <= its final rank computed on the whole dataset，so we can safely discard rows with partial rank > k, anywhere.

This PR adds a per-window-group limit before and after the shuffle to reduce the input data of window processing.

More specifically, the before-shuffle per-window-group limit:

1. adds some row comparison to determine window group boundaries and rank values.

2. applies per-window-group limit to reduce the data size of shuffle, and all the downstream operators.

This is beneficial, assuming the per-window-group data size is large. Otherwise, the method to determine window group boundaries is pure overhead. The config `spark.sql.window.group.limit.threshold` could avoid the overhead if the per-window-group data size is small enough.

The after-shuffle per-window-group limit just applies a per-window-group limit to reduce the data size of window processing. This is very cheap as it only needs to iterate the sorted data (window operator needs to sort the input) once and do some row comparison to determine group boundaries and rank values. It's more efficient to merge it into the window operator, but probably doesn't worth it as the overhead is small.

This PR also take over some functions from https://github.com/apache/spark/pull/34367.


### Why are the changes needed?

1. reduce the shuffle write
2. solve skewed-window problem.
3. improve the performance and TPC-DS.

### Does this PR introduce _any_ user-facing change?
'No'.
Just update the inner implementation and add a new config.


### How was this patch tested?

1. new test suites
2. new micro benchmark
```
Java HotSpot(TM) 64-Bit Server VM 1.8.0_311-b11 on Mac OS X 10.16
Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark Top-K:                                                 Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
-----------------------------------------------------------------------------------------------------------------------------------------------
ROW_NUMBER (PARTITION: , WindowGroupLimit: false)                        10711          11068         363          2.0         510.8       1.0X
ROW_NUMBER (PARTITION: , WindowGroupLimit: true)                          2611           2929         210          8.0         124.5       4.1X
ROW_NUMBER (PARTITION: PARTITION BY b, WindowGroupLimit: false)          23829          24085         178          0.9        1136.2       0.4X
ROW_NUMBER (PARTITION: PARTITION BY b, WindowGroupLimit: true)            6381           6515          97          3.3         304.3       1.7X
RANK (PARTITION: , WindowGroupLimit: false)                              11459          11798         223          1.8         546.4       0.9X
RANK (PARTITION: , WindowGroupLimit: true)                                2588           2837         162          8.1         123.4       4.1X
RANK (PARTITION: PARTITION BY b, WindowGroupLimit: false)                24560          24707         108          0.9        1171.1       0.4X
RANK (PARTITION: PARTITION BY b, WindowGroupLimit: true)                  6395           6530         109          3.3         304.9       1.7X
DENSE_RANK (PARTITION: , WindowGroupLimit: false)                        11563          11740         124          1.8         551.3       0.9X
DENSE_RANK (PARTITION: , WindowGroupLimit: true)                          2563           2819         168          8.2         122.2       4.2X
DENSE_RANK (PARTITION: PARTITION BY b, WindowGroupLimit: false)          24482          24578          64          0.9        1167.4       0.4X
DENSE_RANK (PARTITION: PARTITION BY b, WindowGroupLimit: true)            6368           6511         122          3.3         303.7       1.7X
```

3. manual test on TPC-DS
TPC-DS data size: 2TB.
This improvement is valid for TPC-DS q67 and no regression for other test cases.

| TPC-DS Query   | Before(Seconds)  | After(Seconds)  | Speedup(Percent)  |
|  ----  | ----  | ----  | ----  |
| q44 | 29.945 | 28.862 | 3.75% |
| q67 | 997.7585 | 412.046 | 142.15% |
| q70 | 20.149 | 19.955 | 0.97% |
| All TPC-DS | 7076.6715 | 6472.0535 | 9.34% |

There is an extreme case for this optimization. If the window group size is small enough, we still sort by partition key, order by key before shuffle. The worst case scenario is that there is only one row of data in each window group. I changed the partition of the window frame in q67 of TPC-DS from i_ The category is replaced by the store_ Primary key ss of sales table_ ticket_ Number, you can test this worst-case scenario.

| TPC-DS Query   | Before(Seconds)  | After(Seconds)  | Speedup(Percent)  |
|  ----  | ----  | ----  | ----  |
| q67-changed | 2655.285 | 3062.245 | -13.29% |