@zhengruifeng 2022-11-29
+1 on add a separate config as the threshold to trigger this optimization.

since new logical and physical plans are added, should also consider how it affect existing rules.

thanks for taking this over @beliefer