@cloud-fan 2022-11-29
> since new logical and physical plans are added, should also consider how it affect existing rules.

How about we do this optimization as a planner rule? then logical plan won't change.