@beliefer 2022-12-01
> I just image a special case: if the child's (e.g a SMJ) output ordering happens to satisfy the requirement of this filter, then there will be no extra SortExec, is it always beneficial to apply this optimization?

`WindowGroupLimitExec` need the child 's output ordering like `WindowExec`.