@beliefer 2023-02-03
> Does this one have a chance to Spark 3.4?

Maybe. I testing the performance now.