@cloud-fan 2023-02-08
there are 3 tpc queries having plan change, what's their perf result?