@beliefer 2023-02-09
> there are 3 tpc queries having plan change, what's their perf result?

![image](https://user-images.githubusercontent.com/8486025/217709318-ce005732-48ef-4668-baa7-14ded7faf52f.png)